backend: Move chain related code to sync.go
This moves code related to chain syncing and reorging to a separate
file to ease maintenance.
